+++ cp-library-rs/src/number_theory/comb_no_mod.rs
@@ -1,7 +1,7 @@
 //! 愚直な2項係数の計算
 
 /// $`\binom{n}{r}`$ の値を愚直に求める
-pub fn comb(n: usize, r: usize) -> usize {
+pub fn comb(n: u64, r: u64) -> u64 {
     if r == 0 {
         1
     } else if n < r {
