+++ cp-library-rs/src/number_theory/factorize.rs
@@ -2,7 +2,7 @@
 
 /// 非負整数 $`n`$ を素因数分解し、`素因数`のベクタを返す
 /// - 計算量 : $`O(\sqrt{n})`$
-pub fn factorize_vec(mut n: usize) -> Vec<usize> {
+pub fn factorize_vec(mut n: u64) -> Vec<u64> {
     let mut res = Vec::new();
     for i in 2.. {
         if i * i > n {
@@ -21,7 +21,7 @@ pub fn factorize_vec(mut n: usize) -> Vec<usize> {
 
 /// 非負整数 $`n`$ を素因数分解し、`(素因数,指数)`のベクタを返す
 /// - 計算量 : $`O(\sqrt{n})`$
-pub fn factorize_pairs(mut n: usize) -> Vec<(usize, usize)> {
+pub fn factorize_pairs(mut n: u64) -> Vec<(u64, u64)> {
     let mut res = Vec::new();
     for i in 2.. {
         if i * i > n {
