+++ cp-library-rs/tests/test_crt.rs
@@ -4,15 +4,15 @@ use cp_library_rs::number_theory::crt::garner_algorithm;
 use itertools::Itertools;
 use rand::{rng, Rng};
 
-const P: [usize; 9] = [4, 5, 7, 9, 11, 13, 17, 19, 23];
-const M998: usize = 998244353;
-const M107: usize = 1000000007;
+const P: [u64; 9] = [4, 5, 7, 9, 11, 13, 17, 19, 23];
+const M998: u64 = 998244353;
+const M107: u64 = 1000000007;
 
 #[test]
 fn test_crt_small() {
     let mut rng = rng();
 
-    let prod: usize = P.iter().product();
+    let prod: u64 = P.iter().product();
 
     for _ in 0..5_000 {
         let x = rng.random_range(0..prod);
